04_Mouse_Anomalies/requirements.txt
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
colorama==0.4.6
colorlog==6.9.0
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
fonttools==4.59.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
imbalanced-learn==0.13.0
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
joblib==1.3.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
matplotlib==3.8.4
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
omegaconf==2.3.0
packaging==25.0
pandas==2.2.2
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.1
scikit-learn==1.4.2
scipy==1.11.4
six==1.17.0
sklearn-compat==0.1.3
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
wcwidth==0.2.13
xgboost==1.7.6
